fix master

+++ app/src/main/java/com/devvailonge/flip/features/flashcard/create/presentation/FlashCardCreateEvent.kt
@@ -2,5 +2,5 @@ package com.devvailonge.flip.features.flashcard.create.presentation
 
 sealed class FlashCardCreateEvent {
 
-    data class Insert(val textFront:String, val textBack:String,val categoryId: Int): FlashCardCreateEvent()
+    data class Insert(val textFront:String, val textBack:String,val categoryId: Long): FlashCardCreateEvent()
 }
